refactor: rpc traits
